Use from_array in StructContainer.sync_from_device
Rather than simply passing the array representation of each contained
object to the object class constructor directly, use the object class'
`from_array` method to reconstruct the object.  This provides more
flexibility in object to array (de)serialization.

Update struct container test case to add a `numpy` array to the `Foo`
class.  The `as_array` and `from_array` methods of the `Foo` class
demonstrate the API required for objects to be stored in a
`StructContainer`.